Moving stuff around to comply with the one and only overlord: SonarQube
