+++ Domain/src/main/java/com/infosupport/t2c3/domain/orders/Order.java
@@ -1,7 +1,6 @@
 package com.infosupport.t2c3.domain.orders;
 
 import com.infosupport.t2c3.domain.abs.AbsEntity;
-import com.infosupport.t2c3.domain.customers.CustomerData;
 import java.math.BigDecimal;
 import java.util.List;
 import javax.persistence.*;
